More work on revamping validation in tests

+++ src/main/kotlin/io/craigmiller160/expensetrackerapi/web/types/UpdateTransactionDetailsRequest.kt
@@ -3,6 +3,7 @@ package io.craigmiller160.expensetrackerapi.web.types
 import io.craigmiller160.expensetrackerapi.common.data.typedid.TypedId
 import io.craigmiller160.expensetrackerapi.common.data.typedid.ids.CategoryId
 import io.craigmiller160.expensetrackerapi.common.data.typedid.ids.TransactionId
+import java.math.BigDecimal
 import java.time.LocalDate
 
 data class UpdateTransactionDetailsRequest(
@@ -10,6 +11,6 @@ data class UpdateTransactionDetailsRequest(
   override val confirmed: Boolean,
   val expenseDate: LocalDate,
   val description: String,
-  val amount: Double,
+  val amount: BigDecimal,
   override val categoryId: TypedId<CategoryId>?
 ) : TransactionAndCategoryUpdateItem, TransactionAndConfirmUpdateItem
